Add ShowPastItems component and add to MySchedule.
